# feat: Comprehensive Genesis CLI SDLC Framework Implementation