--- a/cli/services/__init__.py
+++ b/cli/services/__init__.py
@@ -1,0 +1,22 @@
+"""
+Genesis CLI Services Layer
+Service layer architecture following SOLID-CLOUD principles for CLI backend implementation.
+"""
+
+from .auth_service import AuthService
+from .cache_service import CacheService
+from .config_service import ConfigService
+from .error_service import ErrorService
+from .gcp_service import GCPService
+from .performance_service import PerformanceService
+from .terraform_service import TerraformService
+
+__all__ = [
+    "AuthService",
+    "CacheService",
+    "ConfigService",
+    "ErrorService",
+    "GCPService",
+    "PerformanceService",
+    "TerraformService",
+]