--- a/cli/tests/__init__.py
+++ b/cli/tests/__init__.py
@@ -1,0 +1,4 @@
+"""
+Genesis CLI Test Suite
+Comprehensive testing framework for CLI backend implementation following CRAFT methodology.
+"""