… .github/workflows/genesis-cli-comprehensive-testing.yml
… COMPREHENSIVE_SECURITY_ASSESSMENT_REPORT.md
… GENESIS_CLI_TEST_REPORT.md
… GENESIS_HOW_TO_GUIDE.md
… cli/CRAFT_VALIDATION_REPORT.md
… cli/commands/enhanced_infrastructure_commands.py
… cli/completion/bash_completion.sh
… cli/completion/zsh_completion.zsh
… cli/services/auth_service.py
… cli/services/cache_service.py
… cli/services/config_service.py
… cli/services/error_service.py
… cli/services/gcp_service.py
… cli/services/performance_service.py
… cli/services/terraform_service.py
… cli/tests/conftest.py
… cli/tests/test_commands.py
… cli/tests/test_e2e_workflows.py
… cli/tests/test_enhanced_unit_coverage.py
… cli/tests/test_integration_comprehensive.py
… cli/tests/test_performance_optimization.py
… cli/tests/test_services.py
… cli/tests/test_ui_components.py
… cli/ui/__init__.py
… cli/ui/colors.py
… cli/ui/formatter.py
… cli/ui/help.py
… cli/ui/interactive.py
… cli/ui/progress.py
… cli/ui/terminal.py